requirements.txt

telethon
python-dotenv
pytest
PyYAML
requests_oauthlib
requests
firebase_admin
openai
urllib3
schedule
